requirements.txt
psycopg2-binary
pandas
google-cloud-bigquery
python-dotenv
pyyaml
